verificando se arquivo existe

+++ src/utils/file.ts
@@ -1,11 +1,20 @@
-import fs from "fs";
+import fs, { constants } from "fs";
+
+export const fileExists = (fileName: string): boolean => {
+  try {
+    fs.accessSync(fileName, constants.F_OK);
+    return true;
+  } catch (err) {
+    return false;
+  }
+};
 
 export const deleteFile = async (fileName: string): Promise<void> => {
+  if (!fileExists(fileName)) return;
   try {
     await fs.promises.stat(fileName);
+    await fs.promises.unlink(fileName);
   } catch (error) {
-    //
-    return;
+    console.log(error);
   }
-  await fs.promises.unlink(fileName);
 };
